requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
openai
scikit-learn
